LPD-51750 Fix cyclic dependency that cause Cookie.TYPES to be undefined
