+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/util/storage.js
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {CONSENT_TYPES, checkConsent} from './consent';
+import {CONSENT_TYPES, checkConsent} from './cookie/cookie';
 
 export default class Storage {
 	constructor(storage) {
